# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/service": {
            "get": {
                "tags": [
                    "service"
                ],
                "summary": "Returns all the services in list.",
                "description": "Returns all the services saved in database in list.",
                "operationId": "get services",
                "responses": {
                    "200": {
                        "description": "Successfully got the services",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "datas": {
                                            "$ref": "#/components/schemas/Services"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nothing found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not found"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Database error"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "service"
                ],
                "summary": "Store service from post form",
                "description": "Store a new service from form in database, using post method",
                "operationId": "Post new service",
                "responses": {
                    "201": {
                        "description": "Successfully stored in database",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "datas": {
                                            "$ref": "#/components/schemas/Services"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Database error"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/service/{id}": {
            "get": {
                "tags": [
                    "service"
                ],
                "summary": "Returns specified service with details",
                "description": "Return the specified service, by ID, with details",
                "operationId": "get service by Id",
                "parameters": [
                    {
                        "parameter": "get_service_id",
                        "name": "id",
                        "in": "path",
                        "description": "ID of the service",
                        "schema": {
                            "type": "string",
                            "default": "60b8d5d74e00fd5950e78719"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully got the wanted service",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Services"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nothing found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not found"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Database error"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "service"
                ],
                "summary": "Update service from put form based on ID",
                "description": "Update the targeted service from form in database, using post method",
                "operationId": "Update service",
                "parameters": [
                    {
                        "parameter": "get_service_id",
                        "name": "id",
                        "in": "path",
                        "description": "ID of the service",
                        "schema": {
                            "type": "string",
                            "default": "60b927367825c419083d3588"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Succesfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully updated"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Database error"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "service"
                ],
                "summary": "Delete service from Delete method based on ID",
                "description": "Delete the targeted service from form in database, using delete method",
                "operationId": "Delete service",
                "parameters": [
                    {
                        "parameter": "get_service_id",
                        "name": "id",
                        "in": "path",
                        "description": "ID of the service",
                        "schema": {
                            "type": "string",
                            "default": "60b794304e00fd5950e78718"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succesfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully deleted"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ID not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nothing found at this ID"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Database error"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Services": {
                "title": "Service",
                "description": "Service Model",
                "properties": {
                    "_id": {
                        "title": "_id",
                        "description": "_id of the entity",
                        "type": "ObjectId",
                        "example": "60b794304e00fd5950e78718"
                    },
                    "name": {
                        "title": "name",
                        "description": "Name",
                        "type": "string",
                        "example": "café"
                    },
                    "type": {
                        "title": "type",
                        "description": "Type",
                        "type": "integer",
                        "example": "1"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description",
                        "type": "string",
                        "example": "Café expresso de notre cafétaria"
                    },
                    "price": {
                        "title": "price",
                        "description": "Price",
                        "type": "ObjectId",
                        "example": "60b8d5d74e00fd5950e78719"
                    },
                    "startDate": {
                        "title": "Availability start date",
                        "description": "Availability start date",
                        "type": "date",
                        "example": "2021-06-10 00H00"
                    },
                    "endDate": {
                        "title": "Availability end date",
                        "description": "Availability end date",
                        "type": "date",
                        "example": "2021-06-11 00H00"
                    },
                    "memberDiscountAvailable": {
                        "title": "Member availability available",
                        "description": "Member availability available",
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Service"
                }
            }
        }
    },
    "security": [
        []
    ]
}